got detail screen working.
Fixed an issue where we could not get the Jp details.

TODO: Fix delay in loading data.

+++ app/src/main/java/com/buzzed/jpfinder/navigation/JPFinderNavGraph.kt
@@ -1,18 +1,13 @@
 package com.buzzed.jpfinder.navigation
 
-import android.util.Log
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.navigation.NavHostController
-import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import androidx.navigation.navArgument
-import androidx.navigation.navDeepLink
 import com.buzzed.jpfinder.ui.screen.*
 import com.buzzed.jpfinder.ui.screen.ListScreenDestination.communityArg
-import com.buzzed.jpfinder.ui.screen.ListScreenDestination.route
 
 @Composable
 fun JPFinderNavHost(
@@ -54,9 +49,10 @@ fun JPFinderNavHost(
             //deepLinks = listOf(navDeepLink {uriPattern= "uri=android-app://androidx.navigation/list_screen?communityArg={${communityArg}}"})
         ){
             //Log.d("ListScreen","$communityArg     ${it.arguments?.getString("community")}")
+
             DetailsScreen(
                 onNavigateUp = { navController.popBackStack() },
-                id = DetailsScreenDestination.jpId,
+                id = DetailsScreenDestination.jpId.toInt(),
             )
         }
